namespace refactor for build^Celated modules
