+++ docs/source/align.rst
@@ -4,7 +4,7 @@ Align API
 The `Align` module provide a number of functions to translate and orient molecules that is useful during structure
 preparation.
 
-.. automodule:: paprika.align
+.. automodule:: paprika.build.align
    :members:
    :undoc-members:
    :show-inheritance:
